backend/requirements.txt
django
djangorestframework
django_rest_auth
django-allauth
mysqlclient
python-decouple
python-dotenv
django-cors-headers
pycryptodome
requests
celery
django-celery-beat
redis
uvicorn
uvicorn[standard]
channels-redis
boto3